Changed bar behavior so it will not re-open automaticaly after closure
If the bar has been closed on a webpage, it will no longer keep coming
back up every time the page is opened, if it has been manually closed. If
the bar was not closed manually, it will behave normally. A closed bar
can still be accessed via the companion icon.