Moved chrome.pageAction hide/show logic to setPageActionIcon()
Moved the check for the setting to happen along with the regular
setPageActionIcon stuff, since where it was was actually just disabling the bar
along with the icon (whoops).

Added option check for the initial setAllPageActionIcons() call that happens
when the extension loads.

Removed optionChangeHandler and moved logic to existing 'storage' event handler.